[feat][sophia22001] 클린 아키텍처 적용
