+++ src/domain/entities/Todo.ts
@@ -1,0 +1,5 @@
+export interface Todo {
+  id: number;
+  text: string;
+  completed: boolean;
+}
